Rename screen and its related components
