+++ feature/login/src/main/java/com/epmedu/animeal/login/ui/ResendButton.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.epmedu.animeal.base.theme.AnimealTheme
-import com.epmedu.animeal.login.domain.CodeConfirmationViewModel
+import com.epmedu.animeal.login.domain.EnterCodeViewModel
 import com.epmedu.animeal.resources.R
 import java.text.DecimalFormat
 
@@ -70,13 +70,13 @@ private fun ResendButtonPreview() {
             Column(horizontalAlignment = Alignment.CenterHorizontally) {
                 ResendButton(
                     isEnabled = true,
-                    resendDelay = CodeConfirmationViewModel.RESEND_DELAY,
+                    resendDelay = EnterCodeViewModel.RESEND_DELAY,
                     onClick = {}
                 )
                 Divider()
                 ResendButton(
                     isEnabled = false,
-                    resendDelay = CodeConfirmationViewModel.RESEND_DELAY,
+                    resendDelay = EnterCodeViewModel.RESEND_DELAY,
                     onClick = {}
                 )
             }
